Edit resource bug and tests fixed
