--- plugins/entando-plugin-jacms/src/main/java/com/agiletec/plugins/jacms/apsadmin/resource/multipleResource.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE struts PUBLIC
-    "-//Apache Software Foundation//DTD Struts Configuration 2.5//EN"
-    "http://struts.apache.org/dtds/struts-2.5.dtd">
-
-<struts>	
-    <package name="jacms_do/jacms/multipleResource" namespace="/do/jacms/Resource" extends="jacms_do/jacms/Resource">    	
-        
-        <action name="save" class="jacmsMultipleResourceAction" method="save">
-            <result type="tiles" name="input">admin.jacms.resource.entry</result>
-            <result type="tiles" name="failure">admin.jacms.resource.entry</result>
-            <result type="redirectAction">
-                <param name="actionName">list</param>
-                <param name="resourceTypeCode">${resourceTypeCode}</param>
-            </result>
-            <!-- <interceptor-ref name="fileUpload"><param name="maximumSize">10485760</param></interceptor-ref> --><!-- maximun Size = 10 Mb -->
-            <interceptor-ref name="apsFileUpload" />
-            <interceptor-ref name="entandoValidationStack">
-                <param name="requestAuth.requiredPermission">manageResources</param>
-            </interceptor-ref>
-        </action>
-    </package>
-</struts>
