+++ plugins/entando-plugin-jacms/src/main/java/com/agiletec/plugins/jacms/apsadmin/resource/package_en.properties
@@ -88,3 +88,5 @@ error.resource.file.descrEmpty=File description is empty
 message.resource.filename.uploaded=File name  ''{0}'' was uploaded successfully.
 label.add-fileinput=Add Another Resource
 label.remove-fileinput=Remove from form
+error.resource.file.descrTooLong=File description is too long
+error.resource.file.genericError=Error uploading file
