+++ plugins/entando-plugin-jacms/src/main/java/com/agiletec/plugins/jacms/apsadmin/resource/package_it.properties
@@ -90,3 +90,5 @@ error.resource.file.descrEmpty=La descrizione del file \u00e8 vuota
 message.resource.filename.uploaded=Il file ''{0}'' \u00e8 stato caricato con successo.
 label.add-fileinput=Aggiungi una risorsa
 label.remove-fileinput=Rimuovi dal form
+error.resource.file.descrTooLong=La descrizione del file \u00e8 troppo lunga
+error.resource.file.genericError=Errore di caricamento del file
