+++ plugins/entando-plugin-jacms/src/test/java/com/agiletec/plugins/jacms/AllTests.java
@@ -66,7 +66,7 @@ import com.agiletec.plugins.jacms.apsadmin.portal.specialwidget.listviewer.TestN
 import com.agiletec.plugins.jacms.apsadmin.portal.specialwidget.listviewer.TestTextAttributeFilterAction;
 import com.agiletec.plugins.jacms.apsadmin.portal.specialwidget.viewer.TestContentFinderViewerAction;
 import com.agiletec.plugins.jacms.apsadmin.portal.specialwidget.viewer.TestContentViewerWidgetAction;
-import com.agiletec.plugins.jacms.apsadmin.resource.TestResourceAction;
+import com.agiletec.plugins.jacms.apsadmin.resource.TestMultipleResourceAction;
 import com.agiletec.plugins.jacms.apsadmin.resource.TestResourceFinderAction;
 import com.agiletec.plugins.jacms.apsadmin.system.entity.TestJacmsEntityAttributeConfigAction;
 import com.agiletec.plugins.jacms.apsadmin.system.entity.TestJacmsEntityManagersAction;
@@ -187,9 +187,8 @@ public class AllTests {
         suite.addTestSuite(TestPageAction.class);
 
         //Resource
-        suite.addTestSuite(TestResourceAction.class);
+        suite.addTestSuite(TestMultipleResourceAction.class);
         suite.addTestSuite(TestResourceFinderAction.class);
-
         suite.addTestSuite(TestPageActionReferences.class);
 
         //Entity
